+++ packages/terra-hyperlink/CHANGELOG.md
@@ -4,6 +4,9 @@ ChangeLog
 Unreleased
 ----------
 
+### Changed
+* Updated WDIO tests to use themeCombinationOfCustomProperties instead of themeEachCustomProperty
+
 2.8.0 - (April 16, 2019)
 ------------------
 ### Changed