+++ packages/terra-profile-image/CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+### Changed
+* Updated WDIO tests to use themeCombinationOfCustomProperties instead of themeEachCustomProperty
+
 3.7.0 - (April 16, 2019)
 ------------------
 ### Changed